--- blocks/card-cta/card-cta.css
@@ -1,10 +1,0 @@
-/* Block - Card CTA */
-
-
-/* Themes - Card CTA */
-
-
-/* Tablet */
-@media only screen and (min-width: 768px) {
-
-}
