--- blocks/card-cta/card-cta.js
@@ -1,10 +1,0 @@
-export default function decorate(block) {
-  const blockName = block.getAttribute('data-block-name');
-  if (!blockName) {
-    return;
-  }
-
-  [...block.children].forEach((element) => {
-    element.classList.add(`${blockName}__item`);
-  });
-}
