+++ styles/styles.css
@@ -228,6 +228,7 @@
 
   /* box shadow */
   --box-shadow-large: 0 var(--spacer-element-02) var(--spacer-element-06) 0 rgb(185 181 174 / 40%);
+  --box-shadow-small: var(--spacer-element-01) var(--spacer-element-02) var(--spacer-element-02) var(--spacer-element-01) rgba(0 0 0 / 15%);
 }
 
 /* Theme - Global */
